Improve the readme file and delete editorconfig from project.
