+++ components/bl-webcamPhoto-component/.gitignore
@@ -1,1 +1,1 @@
-/dist/!/.editorconfig
+/dist/
