--- components/bl-webcamPhoto-component/src/utils/data-url-to-blob.js
@@ -1,4 +1,0 @@
-export async function dataURLToBlob(dataURL) {
-  const response = await fetch(dataURL);
-  return response.blob();
-}
