--- components/bl-webcamPhoto-component/src/utils/mobile-and-tablet-check.js
@@ -1,14 +1,0 @@
-export const mobileAndTabletCheck = () => {
-  const toMatch = [
-    /Android/i,
-    /webOS/i,
-    /iPhone/i,
-    /iPad/i,
-    /iPod/i,
-    /BlackBerry/i,
-    /Windows Phone/i,
-  ];
-  return toMatch.some(toMatchItem => {
-    return navigator.userAgent.match(toMatchItem);
-  });
-};
