+++ components/bl-webcamPhoto-component/styles/index.less
@@ -189,11 +189,3 @@
     }
   }
 }
-
-
-
-
-
-
-
-
